rename composable names
